+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All the changes, additions, bugfixes, etc. made to this project.
 ## 2022-01-05
 ### Modified
 - Refactor and extracted `store` and `<Provider>` from `index.tsx` into separate files and component, the provider is now called `<ReduxProvider>`
+- `<Task>` now doesn't need `removeTask` as a prop, because it properly uses the `TodoContext`.
 
 ## 2022-01-04
 ### Added